Ajout de nouvelle route pour lister des ressources en fonction de certains paramètre et mise à jour des erreurs

+++ app/routers/user.py
@@ -66,7 +66,9 @@ async def read_user(user_id: int, user: Annotated[str, Depends(decode_token)]):
             "SELECT `id`, `username`, `firstname`, `lastname`, `email`, `rights`, `id_company` FROM user WHERE id=%s", (user_id,))
         result = cursor.fetchone()
         if not result:
-            raise HTTPException(status_code=404, detail="User not found")
+            raise HTTPException(
+                status_code=status.HTTP_404_NOT_FOUND, detail="User not found")
+        # Verify if user is in the same company as the user he wants to see
         if user.id_company != result["id_company"] and user.rights != "MAINTAINER":
             raise HTTPException(
                 status_code=status.HTTP_403_FORBIDDEN,
@@ -75,6 +77,28 @@ async def read_user(user_id: int, user: Annotated[str, Depends(decode_token)]):
         return {"id": result["id"], "username": result["username"], "firstname": f.decrypt(result["firstname"].encode()),
                 "lastname": f.decrypt(result["lastname"].encode()), "email": f.decrypt(result["email"].encode()),
                 "rights": result["rights"], "id_company": result["id_company"]}
+
+
+# TODO: Add a route to get all the users of a company
+@router.get("/users/company/{company_id}")
+async def read_company_users(company_id: int, user: Annotated[str, Depends(decode_token)]):
+    if user.rights != "MAINTAINER":
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="You don't have access to this ressource.",
+        )
+    with connection.cursor() as cursor:
+        cursor.execute(
+            "SELECT `id`, `username`, `firstname`, `lastname`, `email`, `rights`, `id_company` FROM user WHERE id_company=%s", (company_id,))
+        result = cursor.fetchall()
+        if not result:
+            raise HTTPException(
+                status_code=status.HTTP_404_NOT_FOUND, detail="Company not found")
+        for user in result:
+            user["firstname"] = f.decrypt(user["firstname"].encode())
+            user["lastname"] = f.decrypt(user["lastname"].encode())
+            user["email"] = f.decrypt(user["email"].encode())
+        return result
 
 
 # CREATE
@@ -86,12 +110,12 @@ async def create_user(user_create: User,  user: Annotated[str, Depends(decode_to
             detail="You don't have access to this ressource.",
         )
     with connection.cursor() as cursor:
-        cursor.execute("SELECT * FROM user WHERE email=%s",
-                       (f.encrypt(user_create.email.encode()),))
+        cursor.execute("SELECT * FROM user WHERE username=%s",
+                       (user_create.username,))
         result = cursor.fetchone()
         if result:
             raise HTTPException(
-                status_code=409, detail="User already exists")
+                status_code=status.HTTP_409_CONFLICT, detail="User already exists")
         cursor.execute("INSERT INTO user (username, firstname, lastname, password, email, rights, id_company) VALUES (%s, %s, %s, %s, %s, %s, %s)",
                        (user_create.username, f.encrypt(user_create.firstname.encode()), f.encrypt(user_create.lastname.encode()),
                         hashlib.sha256(user_create.password.encode()).hexdigest(
@@ -114,22 +138,37 @@ async def update_user(user_id: int, user_update: User, user: Annotated[str, Depe
         cursor.execute("SELECT * FROM user WHERE id=%s", (user_id,))
         result = cursor.fetchone()
         if not result:
-            raise HTTPException(status_code=404, detail="User not found")
+            raise HTTPException(
+                status_code=status.HTTP_404_NOT_FOUND, detail="User not found")
+        # Verify if user is in the same company as the user he wants to modify
         if user.id_company != result["id_company"] and user.rights != "MAINTAINER":
             raise HTTPException(
                 status_code=status.HTTP_403_FORBIDDEN,
                 detail="You don't have access to this ressource.",
             )
+        # Verify if the user is trying to change the company to another company
         if user.rights != "MAINTAINER" and user.id_company != user_update.id_company:
             raise HTTPException(
                 status_code=status.HTTP_403_FORBIDDEN,
                 detail="You are not allowed to modify the company of this user.",
             )
+        # Verify if user is trying to change a maintainer
+        if user.rights != "MAINTAINER" and result["rights"] == "MAINTAINER":
+            raise HTTPException(
+                status_code=status.HTTP_403_FORBIDDEN,
+                detail="You can't modify a maintainer.",
+            )
+        # Verify if user is trying to change a user to maintainer
+        if user.rights != "MAINTAINER" and user_update.rights == "MAINTAINER":
+            raise HTTPException(
+                status_code=status.HTTP_403_FORBIDDEN,
+                detail="You can't modify rights to maintainer.",
+            )
         cursor.execute("UPDATE user SET username=%s, firstname=%s, lastname=%s, password=%s, email=%s, rights=%s, id_company=%s WHERE id=%s",
                        (user_update.username, f.encrypt(user_update.firstname.encode()), f.encrypt(user_update.lastname.encode()),
-                        hashlib.sha256(
-                            user_update.password.encode()).hexdigest(),
-                        f.encrypt(user_update.email.encode()), user_update.rights, user_update.id_company, user_id))
+                        hashlib.sha256(user_update.password.encode()).hexdigest(
+                       ), f.encrypt(user_update.email.encode()),
+                           user_update.rights, user_update.id_company, user_id))
         connection.commit()
         return {"id": user_id, "username": user_update.username, "firstname": user_update.firstname, "lastname": user_update.lastname,
                 "email": user_update.email, "rights": user_update.rights, "id_company": user_update.id_company}
@@ -147,11 +186,19 @@ async def delete_user(user_id: int, user: Annotated[str, Depends(decode_token)])
         cursor.execute("SELECT * FROM user WHERE id=%s", (user_id,))
         result = cursor.fetchone()
         if not result:
-            raise HTTPException(status_code=404, detail="User not found")
+            raise HTTPException(
+                status_code=status.HTTP_404_NOT_FOUND, detail="User not found")
+        # Verify if user is in the same company as the user he wants to delete
         if user.id_company != result["id_company"] and user.rights != "MAINTAINER":
             raise HTTPException(
                 status_code=status.HTTP_403_FORBIDDEN,
-                detail="You don't have access to this ressource.",
+                detail="You can't delete user from another company.",
+            )
+        # Verify if user is trying to delete a maintainer
+        if user.rights != "MAINTAINER" and result["rights"] == "MAINTAINER":
+            raise HTTPException(
+                status_code=status.HTTP_403_FORBIDDEN,
+                detail="You can't delete a maintainer.",
             )
         cursor.execute("DELETE FROM user WHERE id=%s", (user_id,))
         connection.commit()
